simplify the scoring method

+++ karaoke_view/src/main/java/io/agora/karaoke_view/v11/KaraokeView.java
@@ -36,7 +36,7 @@ public class KaraokeView {
         mScoringAlgorithm = new DefaultScoringAlgorithm();
         mScoringMachine = new ScoringMachine(new VoicePitchChanger(), new ScoringMachine.OnScoringListener() {
             @Override
-            public void onLineFinished(LyricsLineModel line, double score, double cumulativeScore, double perfectScore, int index, int numberOfLines) {
+            public void onLineFinished(LyricsLineModel line, int score, int cumulativeScore, int perfectScore, int index, int numberOfLines) {
                 Log.d(TAG, "onLineFinished " + line + " " + score + " " + cumulativeScore + " " + perfectScore + " " + index + " " + numberOfLines);
 
                 if (mScoringView != null) {
